*Fix sound issue due to bad resample rate *Fix FIR filter apply

+++ src/backend/demodulator/CAm.h
@@ -21,7 +21,7 @@ private:
     // Filters
     CWindowFunc *winfunc;
     // FIR bandpass filter
-    CFIR<int16_t> *filter;
+    CFIR<double> *filter;
 };
 
 #endif // CAM_H
